% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-summary-MEMWAS.R
\name{print.summary.MEMWAS}
\alias{print.summary.MEMWAS}
\title{Print a MEMWAS setup summary}
\usage{
\method{print}{summary.MEMWAS}(x, ...)
}
\arguments{
\item{x}{summary.MEMWAS. Summary object to print.}

\item{...}{list. Additional arguments; currently unused.}
}
\value{
invisible. Invisibly returns \code{x}.
}
\description{
Print the summary produced by \code{summary.MEMWAS()}.
}
\examples{
\dontrun{
print(summary(setup))
}
}
