% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-summary-MEMWAS_fit.R
\name{print.summary.MEMWAS_fit}
\alias{print.summary.MEMWAS_fit}
\title{Print a MEMWAS model summary}
\usage{
\method{print}{summary.MEMWAS_fit}(x, ...)
}
\arguments{
\item{x}{summary.MEMWAS_fit. Summary object to print.}

\item{...}{list. Additional arguments; currently unused.}
}
\value{
invisible. Invisibly returns \code{x}.
}
\description{
Print the summary produced by \code{summary.MEMWAS_fit()}.
}
\examples{
\dontrun{
print(summary(fit))
}
}
